backend/requirements.txt
altgraph==0.17.2
boto3==1.35.24
botocore==1.35.24
Django==5.1.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
future==0.18.2
gunicorn==23.0.0
jmespath==1.0.1
joblib==1.4.2
macholib==1.15.2
numpy==1.26.2
packaging==24.2
pymysql==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.15.0
threadpoolctl==3.5.0
urllib3==1.26.20
